- Fixed compilation on Linux - Removed some goto handling due to C++ incompat (annoying)

+++ src/platform/linux/misc.c
@@ -4,8 +4,8 @@
 void SleepMS(u64 miliseconds)
 {
     struct timespec ts = {
-        .tv_sec  = miliseconds / 1000,
-        .tv_nsec = (miliseconds % 1000) * 1000 * 1000,
+        .tv_sec  = (i64)miliseconds / 1000,
+        .tv_nsec = ((i64)miliseconds % 1000) * 1000 * 1000,
     };
 
     nanosleep(&ts, NULL);
